Fix error which plagued me for the last two hours.
It wasn't the recursion; apparently it never was. It was that apparently,
when `type Sentence = List[Word]`, returning `List[Sentence]()` is not the same
as returning `List(List())`.

This makes me want to bang my head on the wall. The worst part is that it took
so very long to debug this, when I had the shell of things in place hours ago!

Live and learn, I guess.